requirements.txt
Django>=2.2,<2.3
django-allauth==0.41.0
django-cacheops==4.2
django-cors-headers==3.2.1
django-debug-toolbar==2.1
django-filter==2.2.0
django-mptt==0.9.1
django-pglocks==1.0.4
django-prometheus==1.1.0
django-rq==2.2.0
django-tables2==2.2.1
django-taggit==1.2.0
django-taggit-serializer==0.1.7
django-timezone-field==4.0
djangorestframework==3.10.3
drf-yasg[validation]==1.17.0
Jinja2==2.10.3
Markdown==2.6.11
netaddr==0.7.19
Pillow==7.0.0
psycopg2-binary==2.8.4
py-gfm==0.1.4
pycryptodome==3.9.4
PyYAML==5.3
redis==3.3.11
svgwrite==1.3.1
